requirements.txt
stockfish
selenium
python-chess
keyboard